fix: move `QuizLayout` to a `UserQuiz` folder
